+++ src/tabpfn/preprocessors/add_fingerprint_features_step.py
@@ -8,7 +8,7 @@ from typing_extensions import override
 import numpy as np
 import torch
 
-from tabpfn.preprocessors.helpers import (
+from tabpfn.preprocessors.preprocessing_helpers import (
     FeaturePreprocessingTransformerStep,
 )
 from tabpfn.utils import infer_random_state
