+++ src/tabpfn/preprocessors/differentiable_z_norm_step.py
@@ -6,7 +6,7 @@ from typing_extensions import override
 
 import torch
 
-from tabpfn.preprocessors.helpers import (
+from tabpfn.preprocessors.preprocessing_helpers import (
     FeaturePreprocessingTransformerStep,
 )
